+++ packages/liveblocks-react/src/room.tsx
@@ -53,9 +53,7 @@ import * as React from "react";
 import { useSyncExternalStoreWithSelector } from "use-sync-external-store/shim/with-selector.js";
 
 import { RoomContext, useIsInsideRoom, useRoomOrNull } from "./contexts";
-import { byFirstCreated, byMostRecentlyUpdated } from "./lib/compare";
 import { isString } from "./lib/guards";
-import { makeThreadsFilter } from "./lib/querying";
 import { retryError } from "./lib/retry-error";
 import { useInitial } from "./lib/use-initial";
 import { useLatest } from "./lib/use-latest";
@@ -65,6 +63,7 @@ import {
   getUmbrellaStoreForClient,
   LiveblocksProviderWithClient,
   selectInboxNotifications,
+  selectThreads,
   useClient,
   useClientOrNull,
 } from "./liveblocks";
@@ -85,7 +84,6 @@ import type {
   RoomNotificationSettingsStateSuccess,
   RoomProviderProps,
   StorageStatusSuccess,
-  ThreadsQuery,
   ThreadsState,
   ThreadsStateSuccess,
   ThreadSubscription,
@@ -166,39 +164,6 @@ function selectorFor_useOthersConnectionIds(
   others: readonly User<JsonObject, BaseUserMeta>[]
 ): number[] {
   return others.map((user) => user.connectionId);
-}
-
-/**
- * @private Do not rely on this internal API.
- */
-// TODO This helper should ideally not have to be exposed at the package level!
-// TODO It's currently used by react-lexical though.
-export function selectThreads<M extends BaseMetadata>(
-  state: UmbrellaStoreState<M>,
-  options: {
-    roomId: string | null;
-    query?: ThreadsQuery<M>;
-    orderBy:
-      | "age" // = default
-      | "last-update";
-  }
-): ThreadData<M>[] {
-  let threads = state.threads;
-
-  if (options.roomId !== null) {
-    threads = threads.filter((thread) => thread.roomId === options.roomId);
-  }
-
-  // Third filter pass: select only threads matching query filter
-  const query = options.query;
-  if (query) {
-    threads = threads.filter(makeThreadsFilter<M>(query));
-  }
-
-  // Sort threads by creation date (oldest first)
-  return threads.sort(
-    options.orderBy === "last-update" ? byMostRecentlyUpdated : byFirstCreated
-  );
 }
 
 function selectNotificationSettings<M extends BaseMetadata>(
